Empty commit, see description
This commit is to push and explain the branch "server".  The server branch will be a stable branch free from experimental changes, the branch I will also be running the bot on.  After things have been tested all changes will be pushed to that branch as well.  This is just so I can make changes to master without having to worry about breaking things as much.