Move stuff back to the main irc class
The code is not specific to twitch.  This change also adds support for
SSL sockets.  It also makes the socket timeout every 10 seconds instead
so that it can check if the loop needs to continue to allow the thread
to be released if all the bots on the connection are removed.